STYLE: Remove `static_cast<unsigned int>` from dimension constants
Dimensions are already `unsigned int`, so these casts had no effect anyway.

Replaced `static_cast<unsigned int>\(([A-Za-z]+Dimension)\)` with `$1`, using
regular expressions.

Replaced `static_cast<unsigned int>(Self::CellDimension)` with `CellDimension`.

+++ Modules/Core/Common/include/itkExtractImageFilter.hxx
@@ -130,7 +130,7 @@ ExtractImageFilter<TInputImage, TOutputImage>::GenerateOutputInformation()
   typename OutputImageType::DirectionType outputDirection;
   typename OutputImageType::PointType     outputOrigin{};
 
-  if (static_cast<unsigned int>(OutputImageDimension) > static_cast<unsigned int>(InputImageDimension))
+  if (OutputImageDimension > InputImageDimension)
   {
     // copy the input to the output and fill the rest of the
     // output with zeros.
